New function GenerateCodeSPSS_WOEVariables()
datamining-functions.r:
- New functions:
	- GenerateCodeSPSS_WOEVariables(): function created from the part of
InformationValue() where the SPSS code for the WOE variables calculation
is generated. The goal is to be able to call a function that reads a WOE
table and generates the SPSS code.
	When doing this, I changed the structure on which the SPSS code is
returned, from a data frame to a vector, since it is more easily
handled.

- Updated functions:
	- InformationValue():
		- now calls the new function GenerateCodeSPSS_WOEVariables() to
generate the SPSS code.
		- there is a new column in the IV and WOE outputs called "vartype",
which indicates whether the variable is "character" or "numeric", which
is important when generating the code for categorical variables since we
need to know whether to enclose the value against which they are
compared in quotes or not.

startup-functions.r
- New functions:
	- getBounds(), getVarType(): functions taken from InformationValue() in
datamining-functions.r and moved here since they are general-purpose
functions.